e/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ai-scraper"
version = "0.1.0"
description = "A modular and extensible web scraping library for AI agents"
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
dependencies = [
    "playwright>=1.40.0",
    "openai>=1.0.0",
    "google-generativeai>=0.3.0",
    "ollama>=0.1.0",
]
requires-python = ">=3.9"

[project.optional-dependencies]
dev = [
    "pytest>=7.0",
    "black>=23.0",
    "isort>=5.0",
    "mypy>=1.0",
]

[tool.black]
line-length = 88
target-version = ["py39"]

[tool.isort]
profile = "black"
multi_line_output = 3 